+++ src/main/java/model/position/Position.java
@@ -1,4 +1,4 @@
-package point;
+package model.position;
 
 import java.util.Objects;
