+++ src/test/java/model/piece/KingTest.java
@@ -1,13 +1,13 @@
-package piece;
+package model.piece;
 
 import java.util.stream.Stream;
 import model.Camp;
+import model.position.Position;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import point.Position;
 
 class KingTest {
